style: Use fabsf() instead of fabs() for float math
COSMETIC FIX: Changed fabs() to fabsf() in plausibility check for consistency.

Why fabs vs fabsf:
- fabs() takes double, less efficient for float math
- fabsf() takes float, preferred for embedded systems
- Rest of codebase consistently uses fabsf()
- No functional difference (implicit float->double->float conversions)

Changes:
- Line 126: fabs(rawDist - mean) → fabsf(rawDist - mean)

Impact: Minor performance improvement, consistency with code style
Severity: COSMETIC (no functional change)

+++ firmware/tank-sensor/src/sensor.cpp
@@ -123,7 +123,7 @@ static float applyTemporalFilter(float rawDist) {
     float mean = sum / (HISTORY_SIZE - 2);
 
     // Plausibility check: reject if raw differs from mean by >2cm
-    if (fabs(rawDist - mean) > 20.0f) {
+    if (fabsf(rawDist - mean) > 20.0f) {
         Serial.printf("[Filter] PLAUSIBLE_REJECT: raw=%.1f cm, mean=%.1f cm (>2cm threshold)\n",
                       rawDist, mean);
         gDiag.lastFilteredDist = mean;
